Remove `Wgpu2d`'s lifetime parameter
This is required for running with `winit` event loop.
`EventLoop::run` requires the event handler to be `'static`,
but the `&'static Device` is almost impossible to get without dirty
hack, so we move it from `Wgpu2d`'s fields to `WgpuGraphics`'s one.
`WgpuGraphics` is normally created inside the event handler, so the
lifetime parameter doesn't need to be `'static`.